[CLEANUP]
- Rolled back AbstractPentahoAsyncExecutor and AbstractJobManager. With proper generics defined, the normal classes will do just fine.
- Moved PentahoAsyncReportExecution to AbstractAsyncReportExecution and made the "getState" method abstract. The generics system does not allow us to return concrete classses for parametrzied types. The  PentahoAsyncReportExecution class now then hard-codes the type of the state so that we can return the concrete implementation from our listener.
- hardcoded the return type for the callable, we always return IFxedSizeStreamingContent. There is no valid reason to have that parametrized. (And hard-coding it makes the code easier to read)
