+++ src/org/pentaho/reporting/platform/plugin/async/PentahoAsyncReportExecution.java
@@ -1,67 +1,30 @@
 package org.pentaho.reporting.platform.plugin.async;
 
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.io.input.NullInputStream;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.pentaho.platform.api.engine.IPentahoSession;
 import org.pentaho.platform.engine.core.audit.AuditHelper;
 import org.pentaho.platform.engine.core.audit.MessageTypes;
 import org.pentaho.platform.engine.core.system.PentahoSessionHolder;
-import org.pentaho.reporting.libraries.base.util.ArgumentNullException;
 import org.pentaho.reporting.platform.plugin.SimpleReportingComponent;
 import org.pentaho.reporting.platform.plugin.staging.AsyncJobFileStagingHandler;
 import org.pentaho.reporting.platform.plugin.staging.IFixedSizeStreamingContent;
 
-import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.UUID;
-import java.util.concurrent.FutureTask;
-import java.util.concurrent.RunnableFuture;
 
-public class PentahoAsyncReportExecution implements IAsyncReportExecution<IFixedSizeStreamingContent, IAsyncReportState> {
-
-  protected final SimpleReportingComponent reportComponent;
-  protected final AsyncJobFileStagingHandler handler;
-  protected final String url;
-  protected final String auditId;
-  protected final IPentahoSession safeSession;
-
-  private AsyncReportStatusListener listener;
+public class PentahoAsyncReportExecution extends AbstractAsyncReportExecution<IAsyncReportState> {
 
   private static final Log log = LogFactory.getLog( PentahoAsyncReportExecution.class );
 
-  /**
-   * Creates callable execuiton task.
-   *
-   * @param url             for audit purposes
-   * @param reportComponent component responsible for gererating report
-   * @param handler         content staging handler between requests
-   */
-  public PentahoAsyncReportExecution( final String url,
-                                      final SimpleReportingComponent reportComponent,
-                                      final AsyncJobFileStagingHandler handler,
-                                      final IPentahoSession safeSession,
-                                      final String auditId ) {
-    this.reportComponent = reportComponent;
-    this.handler = handler;
-    this.url = url;
-    this.auditId = auditId;
-    this.safeSession = safeSession;
+  public PentahoAsyncReportExecution(String url,
+                                     SimpleReportingComponent reportComponent,
+                                     AsyncJobFileStagingHandler handler,
+                                     IPentahoSession safeSession,
+                                     String auditId) {
+    super(url, reportComponent, handler, safeSession, auditId);
   }
 
-  public void notifyTaskQueued( final UUID id ) {
-    ArgumentNullException.validate( "id", id );
-
-    if ( listener != null ) {
-      throw new IllegalStateException( "This instance has already been scheduled." );
-    }
-    this.listener = createListener( id );
-  }
-
-  protected AsyncReportStatusListener getListener() {
-    return this.listener;
-  }
 
   /**
    * Generate report and return input stream to a generated report from server.
@@ -74,6 +37,7 @@ public class PentahoAsyncReportExecution implements IAsyncReportExecution<IFixed
    * @throws Exception
    */
   @Override public IFixedSizeStreamingContent call() throws Exception {
+    final AsyncReportStatusListener listener = getListener();
     if ( listener == null ) {
       throw new NullPointerException( "No listener for async report execution: " + url );
     }
@@ -86,13 +50,13 @@ public class PentahoAsyncReportExecution implements IAsyncReportExecution<IFixed
       ReportListenerThreadHolder.setListener( listener );
       final long start = System.currentTimeMillis();
       AuditHelper.audit( safeSession.getId(), safeSession.getId(), url, getClass().getName(), getClass().getName(),
-          MessageTypes.INSTANCE_START, auditId, "", 0, null );
+              MessageTypes.INSTANCE_START, auditId, "", 0, null );
 
       if ( reportComponent.execute() ) {
 
         final long end = System.currentTimeMillis();
         AuditHelper.audit( safeSession.getId(), safeSession.getId(), url, getClass().getName(), getClass().getName(),
-            MessageTypes.INSTANCE_END, auditId, "", ( (float) ( end - start ) / 1000 ), null );
+                MessageTypes.INSTANCE_END, auditId, "", ( (float) ( end - start ) / 1000 ), null );
 
         listener.setStatus( AsyncExecutionStatus.FINISHED );
 
@@ -123,75 +87,13 @@ public class PentahoAsyncReportExecution implements IAsyncReportExecution<IFixed
     }
   }
 
-  protected void fail() {
-    // do not erase canceled status
-    if ( listener != null && !listener.getState().getStatus().equals( AsyncExecutionStatus.CANCELED ) ) {
-      listener.setStatus( AsyncExecutionStatus.FAILED );
-    }
-    AuditHelper.audit( safeSession.getId(), safeSession.getId(), url, getClass().getName(), getClass().getName(),
-        MessageTypes.FAILED, auditId, "", 0, null );
-  }
-
-  protected AsyncReportStatusListener createListener( final UUID instanceId ) {
-    return new AsyncReportStatusListener( getReportPath(), instanceId, getMimeType() );
-  }
-
-  // cancel can only be called from the future created by "newTask".
-  protected void cancel() {
-    String userName = safeSession == null ? "Unknown" : safeSession.getName();
-    log.info( "Report execution canceled: " + url + " , requested by : " + userName );
-    this.listener.setStatus( AsyncExecutionStatus.CANCELED );
-  }
-
-  @Override public RunnableFuture<IFixedSizeStreamingContent> newTask() {
-    return new FutureTask<IFixedSizeStreamingContent>( this ) {
-      public boolean cancel( final boolean mayInterruptIfRunning ) {
-        try {
-          if ( mayInterruptIfRunning ) {
-            PentahoAsyncReportExecution.this.cancel();
-          }
-        } catch ( final Exception e ) {
-          // ignored.
-        }
-        return super.cancel( mayInterruptIfRunning );
-      }
-    };
-  }
-
-  @Override public void requestPage( final int page ) {
-    listener.setRequestedPage( page );
-  }
-
-  @Override public String toString() {
-    return "PentahoAsyncReportExecution{" + "url='" + url + '\'' + ", instanceId='" + auditId + '\'' + ", listener="
-        + listener + '}';
-  }
-
-  @Override public IAsyncReportState getState() {
+  @Override
+  public IAsyncReportState getState() {
+    final AsyncReportStatusListener listener = getListener();
     if ( listener == null ) {
       throw new IllegalStateException( "Cannot query state until job is added to the executor." );
     }
     return listener.getState();
   }
 
-  public String getMimeType() {
-    return reportComponent.getMimeType();
-  }
-
-  public String getReportPath() {
-    return this.url;
-  }
-
-  public static final IFixedSizeStreamingContent NULL = new NullSizeStreamingContent();
-
-  public static final class NullSizeStreamingContent implements IFixedSizeStreamingContent {
-
-    @Override public InputStream getStream() {
-      return new NullInputStream( 0 );
-    }
-
-    @Override public long getContentSize() {
-      return 0;
-    }
-  }
 }
